requirements.txt
flask
python-dotenv
openai
sendgrid
pydub
